Factored out clone options from BitBucket translator
