+++ app/lib/bk/compat/parsers/bitbucket/steps.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require_relative '../../pipeline/step'
-require_relative '../../pipeline/plugin'
 
 require_relative 'caches'
+require_relative 'clone'
 require_relative 'image'
 require_relative 'services'
 require_relative 'shared'
@@ -95,32 +95,6 @@ module BK
             'BITBUCKET_STEP_OIDC_TOKEN="$(buildkite-agent oidc request-token)"',
             'export BITBUCKET_STEP_OIDC_TOKEN'
           ]
-        end
-
-        def translate_clone(opts)
-          sparse_checkout = opts.delete('sparse-checkout')
-          enabled = opts.delete('enabled')
-          msg = '# repository interactions (clone options) should be configured in the agent itself'
-
-          # nothing specified
-          return [] if sparse_checkout.nil? && enabled.nil? && opts.empty?
-
-          BK::Compat::CommandStep.new(commands: msg).tap do |cmd|
-            cmd.env['BUILKITE_REPO'] = '' unless enabled.nil? || enabled
-            cmd << sparse_checkout_plugin(sparse_checkout)
-          end
-        end
-
-        def sparse_checkout_plugin(conf)
-          return [] if conf.nil? || !conf.fetch('enabled', true)
-
-          BK::Compat::Plugin.new(
-            name: 'sparse-checkout',
-            config: {
-              'paths' => conf['patterns'],
-              'no-cone' => !conf.fetch('cone-mode', true)
-            }
-          )
         end
 
         def translate_agents(conf)
